% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_package.R
\name{build_package}
\alias{build_package}
\title{Build package}
\usage{
build_package()
}
\arguments{
\item{package_name}{Package to document,install and load}
}
\value{
None
}
\description{
Build package
}
\examples{
build_package(nilsPaket)
}
