% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonly_mut_genes.R
\name{commonly_mut_genes}
\alias{commonly_mut_genes}
\title{commonly_mut_genes}
\usage{
commonly_mut_genes(ClusterDF)
}
\arguments{
\item{ClusterDF}{ClusterDF}
}
\value{
Table with which genes in which cluster
}
\description{
Find genes that are commonly mutated from ClusterDF
}
\examples{
commonly_mut_genes(ClusterDF)
}
