% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cancer_type.R
\name{get_cancer_type}
\alias{get_cancer_type}
\title{get_cancer_type}
\usage{
get_cancer_type(TCGA_code)
}
\arguments{
\item{samples}{Samples one wants to find cancer type of.}
}
\value{
TSS_code
}
\description{
Function that given a vector of samples return a vector of their cancer type
}
\examples{
get_cancer_type("TCGA-BT-A2LB-01A" "TCGA-DK-A1AC-01A" "TCGA-DK-A3WW-01A")
}
