% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contributing_signatures.R
\name{get_contributing_signatures}
\alias{get_contributing_signatures}
\title{get_contributing_signatures}
\usage{
get_contributing_signatures(ClusterDF, cos_sim_samples_cosmic,
  mutational_matrix, threshold)
}
\arguments{
\item{ClusterDF}{Dataframe with cluster information}

\item{mutational_matrix}{Mutational matrix from mut_mat()}

\item{cos_sim_matrix}{cosine simililarity matrix}

\item{treshold}{Treshold for needed cosine similarity between reconstructed and original mutational matrix}
}
\value{
DF with main contributing signatures in each cluster
}
\description{
Function that finds the minimun number of signatures needed recreated mutational profiles of each cluster to a certain threshold
However cant deal with clusters of size = 1. These are removed.
}
\examples{
get_signatre(Cluster,cos_sim_matrix,mutational_matrix,treshold=0.95)
}
