% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exp_as_matrix.R
\name{get_exp_as_matrix}
\alias{get_exp_as_matrix}
\title{get_exp_as_matrix}
\usage{
get_exp_as_matrix(samples, genes, DF)
}
\arguments{
\item{samples}{vector of samples}

\item{genes}{vector of genes}

\item{DF}{DF from mRNA_exp function}
}
\value{
mRNA_exp.m
}
\description{
Function that given vector of samples and genes and a mRNA expression table(from mRNA_exp function) returns a matrix with exp/avg exp
}
\examples{
get_exp_as_matrix(samples,genes,DF)
}
