% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_high_CNV_samples.R
\name{get_high_CNV_samples}
\alias{get_high_CNV_samples}
\title{Extract high mutation samples}
\usage{
get_high_CNV_samples(threshold)
}
\arguments{
\item{treshold}{What top percentile should be taken from each cancer type}
}
\value{
high_cnv
}
\description{
Extract high mutation samples
}
\examples{
read_GISTIC(MC3,type="SD",3)
}
