% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_high_mutations.R
\name{get_high_mutations}
\alias{get_high_mutations}
\title{Extract high mutation samples}
\usage{
get_high_mutations(MC3_file, type, treshold)
}
\arguments{
\item{type}{sets treshold type , either "absolute" or "SD"}

\item{treshold}{numerical treshold value either absolute number or number of SD:s}

\item{MC3_files}{MC3 data frame created by read_MC3}
}
\value{
None
}
\description{
Extract high mutation samples
}
\examples{
read_GISTIC(MC3,type="SD",3)
}
