% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutational_matrix.R
\name{get_mutational_matrix}
\alias{get_mutational_matrix}
\title{get_mutational_matrix}
\usage{
get_mutational_matrix(VCF_files, file_name)
}
\arguments{
\item{VCF_files}{List of directories to VCF files to extract a mutational matrix from}

\item{file_name}{Name of file you want to createBothm mutmat and GRange object.}
}
\description{
Function that given vector of samples and genes and a mRNA expression table(from mRNA_exp function) returns a matrix with exp/avg exp
}
\examples{
get_mutational_matrix(VCF_files,"Run3")
}
