% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature.R
\name{get_signature}
\alias{get_signature}
\title{get_signature}
\usage{
get_signature(ClusterDF, cos_sim_matrix, treshold)
}
\arguments{
\item{cos_sim_matrix}{}

\item{Cluster}{}

\item{Treshold}{Treshold for colmean over cos_sim_similarity}
}
\value{
DF with signatures and their colmean of cosSim
}
\description{
Function that return what clusters contain what signatures(takes mean)
}
\examples{
get_signatre(Cluster,cos_sim_matrix,Treshold=0.7)
}
