% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_mut.R
\name{high_mut}
\alias{high_mut}
\title{Extract high mutation samples}
\usage{
high_mut(treshold)
}
\arguments{
\item{treshold}{numerical treshold value setting number of standard deviations}
}
\value{
None
}
\description{
Extract high mutation samples
}
\examples{
high_mut(3)
}
