% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNA_exp.R
\name{mRNA_exp}
\alias{mRNA_exp}
\title{mRNA_exp}
\usage{
mRNA_exp(genes)
}
\arguments{
\item{genes}{Genes to make a subselection from (to large otherwise)}
}
\value{
mRNA_DF
}
\description{
Given some genes get mRNA expression for all samples in TCGA cohort
}
\examples{
mRNA_exp(c("ACC","XX","YY"))
}
