% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart_cancer_cluster.R
\name{piechart_cancer_cluster}
\alias{piechart_cancer_cluster}
\title{piechart_cancer_cluster}
\usage{
piechart_cancer_cluster(ClusterDF, mfrows = c(3, 2))
}
\arguments{
\item{ClusterDF}{blahblah}

\item{mfrows}{Subplot layout(depends on cluster)}
}
\description{
Plot piechart over cluster cancer dist
}
\examples{
piechart_cancer_cluster(DF,c(2,3))
}
