% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster.R
\name{plot_cluster}
\alias{plot_cluster}
\title{plot_cluster}
\usage{
plot_cluster(cluster, N, y_pos)
}
\arguments{
\item{cluster}{Cluster to cut and visualise}

\item{N}{Number of clusters to be created}

\item{y_pos}{Define ypos where cluster identity is plotted (2.2)}
}
\description{
plot_cluster in order to visualise cut-tree
}
\examples{
plot_cluster(cluster,7,2.2)
}
