% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_in_cosine.R
\name{plot_cluster_in_cosine}
\alias{plot_cluster_in_cosine}
\title{plot_cluster_in_cosine}
\usage{
plot_cluster_in_cosine(cluster, cos_sim_matrix, N)
}
\arguments{
\item{cluster}{Clustered samples from hclust}

\item{cos_sim_matrix}{cosine_similarity matrix from cos_sim_matrix(mutational patterns)}

\item{N}{Number of clusters to create from cutree.}
}
\value{
DF with cluster identity for each sample
}
\description{
Plot cosine similiarity with clustered samples
}
\examples{
plot_cluster_in_cosine(cluster,cos_sim_matrix,8)
}
