% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mRNA_SNV.R
\name{plot_mRNA_SNV}
\alias{plot_mRNA_SNV}
\title{plot_mRNA_SNV}
\usage{
plot_mRNA_SNV(samples, genes, cluster_names = FALSE, mRNA_DF)
}
\arguments{
\item{samples}{List of sample-id}

\item{genes}{List of Hugo- gene id:s}

\item{cluster_names}{Plot cluster id on left side? False by default}

\item{mRNA_DF}{Data frame containing mRNA data( from mRNA_exp(genes))}
}
\description{
Function that given a list of genes and samples plot the CNV/SNV correlation heatmap
}
\examples{
plot_mRNA_SNV(c(XX,YY),c("ACC","BRCA1"),"Good title",cluster_names=FALSE,mRNA_DF)
}
