% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_GISTIC.R
\name{read_GISTIC}
\alias{read_GISTIC}
\title{Read GISTIC file}
\usage{
read_GISTIC(files2read, selectProteins)
}
\arguments{
\item{files2read}{Vector of files to read}

\item{selectProteins}{Vector of proteins names to only select from Defaults to TRUE}
}
\value{
None
}
\description{
Read Gistic file
}
\examples{
read_GISTIC(files,DNArepair_proteins)
}
