% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MC3.R
\name{read_MC3}
\alias{read_MC3}
\title{Read MC3 file}
\usage{
read_MC3(genes2select)
}
\arguments{
\item{selectProteins}{Vector of genenames to make a subselction from}
}
\value{
MC3_DF
}
\description{
Read MC3 file
}
\examples{
read_MC3(DNArepair_genes)
}
