% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CNV_SNV.R
\name{plot_CNV_SNV}
\alias{plot_CNV_SNV}
\title{plot_CNV_SNV}
\usage{
plot_CNV_SNV(samples, genes)
}
\arguments{
\item{samples}{List of sample-id}

\item{genes}{List of Hugo- gene id:s}
}
\description{
Function that given a list of genes and samples plot the CNV/SNV correlation heatmap
}
\examples{
plot_CNV_SNV(c(XX,YY),c("ACC","BRCA1"))
}
